03-10-2016  Pequeños cambios: - Validacion de vista de clientes. - unificacion de vista de catalogo - Programacion de creacion de catalogo

+++ application/views/pages/menu.php
@@ -9,18 +9,43 @@
             </div>
         </header>
        <div id="menu">
-           <ul class="nav menu demo-navigation mdl-navigation__link" >
-               <a href="Main"><li><i class="material-icons">home</i> inicio</li></a>
-               <a href="EliminarVineta"><li><i class="material-icons">remove_circle</i> eliminar boucher</li></a>
-               <a href="Clientes"><li><i class="material-icons">supervisor_account</i> clientes</li></a>
-               <!--<a href="BajaClientes"><li><img src="<?PHP echo base_url()?>assets/img/bajacliente.png" width="30px"> baja clientes</li></a>-->
-               <a href="PuntosClientes"><li><i class="material-icons">content_copy</i> puntos clientes</li></a>
-               <a href="Frp"><li><i class="material-icons">payment</i> canje puntos (frp)</li></a>
-               <a href="Catalogo"><li><i class="material-icons">dashboard</i> catálogo</li></a>
-               <a href="Usuarios"><li><i class="material-icons">account_box</i> usuarios</li></a>
-               <a href="Reportes"><li><i class="material-icons">description</i> reportes</li></a>
-               <a href="salir"> <li><i class="material-icons">exit_to_app</i> cerrar sesión</li></a>
-          </ul>
+          <?php if ($this->session->userdata('IdRol')==1){
+            echo '<ul class="nav menu demo-navigation mdl-navigation__link" >
+                    <a href="Main"><li><i class="material-icons">home</i> inicio</li></a>
+                    <a href="EliminarVineta"><li><i class="material-icons">remove_circle</i> eliminar boucher</li></a>
+                    <a href="Clientes"><li><i class="material-icons">supervisor_account</i> clientes</li></a>
+                    <a href="PuntosClientes"><li><i class="material-icons">content_copy</i> puntos clientes</li></a>
+                    <a href="Frp"><li><i class="material-icons">payment</i> canje puntos (frp)</li></a>
+                    <a href="Catalogo"><li><i class="material-icons">dashboard</i> catálogo</li></a>
+                    <a href="Usuarios"><li><i class="material-icons">account_box</i> usuarios</li></a>
+                    <a href="Reportes"><li><i class="material-icons">description</i> reportes</li></a>
+                    <a href="salir"> <li><i class="material-icons">exit_to_app</i> cerrar sesión</li></a>                  
+                  </ul>';
+            }
+            if ($this->session->userdata('IdRol')==2){
+            echo '<ul class="nav menu demo-navigation mdl-navigation__link" >
+                    <a href="Main"><li><i class="material-icons">home</i> inicio</li></a>
+                    <a href="Clientes"><li><i class="material-icons">supervisor_account</i> clientes</li></a>
+                    <div id="sac">
+                  <img src="'.base_url().'assets/img/sac_atencion.png" alt="Icon" />
+                    </div>
+                  </ul>';
+            echo '<ul id="conTIcon2" class="nav menu demo-navigation mdl-navigation__link" >                    
+                    <a href="salir"><li><i class="material-icons">exit_to_app</i> cerrar sesión</li></a>
+                  </ul>';
+            }
+            else if ($this->session->userdata('IdRol')==3){
+            echo '<ul class="nav menu demo-navigation mdl-navigation__link" >
+                    <a href="Main"><li><i class="material-icons">home</i> inicio</li></a>
+                    <a href="Clientes"><li><i class="material-icons">supervisor_account</i> clientes</li></a>
+                    <div id="sac">
+                  <img src="'.base_url().'assets/img/sac_atencion.png" alt="Icon" />
+                    </div>
+                  </ul>';
+            echo '<ul id="conTIcon2" class="nav menu demo-navigation mdl-navigation__link" >                    
+                    <a href="salir"><li><i class="material-icons">exit_to_app</i> cerrar sesión</li></a>
+                  </ul>'; 
+            }
+          ?>         
        </div>
     </div>
-
